Move OutlinedChip component into higher directory
