+++ src/components/projections/Edit/ProjectionList.tsx
@@ -2,11 +2,11 @@ import type { ProjectionListProps } from 'src/components/projections/Edit/types'
 
 import { Box, Stack } from '@mui/material';
 
-import { OutlinedChip } from 'src/components/shared/OutlinedChip';
 import { useUpdateDraftedProjection } from 'src/hooks/projections/useUpdateDraftedProjection';
 import { logRocketEvent } from 'src/services/shared';
 import { CustomEvents } from 'src/services/types';
 import { useWorkflowStore } from 'src/stores/Workflow/Store';
+import { OutlinedChip } from 'src/styledComponents/chips/OutlinedChip';
 
 export const ProjectionList = ({
     collection,
