+++ src/styledComponents/chips/OutlinedChip.tsx
